Morse code
First I check the argument amount, if it's not 2 the program exits. After this I take the first argument and put it to uppercase and check if it contains only alphabets and spaces. Finally I browse through the morse code dictionary and match each letter to the code and print it out.