Add krusthol's solution for String to Morse
Simple Python3 executable script that encodes string to morse, using import sys and import string for checking that arguments supplied equals 1 and that argument is within bounds of a-zA-Z plus space. For printing morse, uses a if else if tree that checks against lowercased characters of argument, and stdout writes a-z to respective morse, letting spaces be spaces. Finishes with a newline and syscode 0