My simple approach to the challenge
Parse the string by testing if it's alphabet or space and then making chars into lowercase, getting the index and returning the matching morse code which gets added to the string.

If something else is parsed it will print out usage and finish the function without printing already parse string.