Added chall02
First what I do in the given function is to check if RBG values are respected to avoid misbehavior (0-255), I first allocate enough space for the string which will be returned in the end (1 hash and 6 hexvalues), and for each RGB value I calculate their Hexadecimal value and strjoin these values to my first string previously allocated. For values from 0 to 9, I decided to use itoa() function on base 10 to convert these values from integers to chars and for the rest (10-15) I simply set their respective Hexadecimal values (a-f).