My solution
First get the response. Make it into a string. Then I parse it to get the id, and color values with a very complicated "while" loop. Then I convert the 3 color values to hexadecimals. I append the strings together to get the color code. Then I combine the base url and the id and color value. Then I make the second request and print out the http response.