chall04
First I get the input from standard input, until control-d is pressed. Then I parse the newlines out and check for bad input. I put spaces at the end of each line if needed so that there is as many characters in each line as the input defined is the width to be. Then I reverse the list so I can pull the sand from the bottom-most row to the bottom first. After checking each character in every line and pulling the sand as far down as possible, I reverse the list again and print it out.