Minecraft physics
Loop input until complete input has been gotten. First get the size by using Atoi. Gives error if not valid
Then start to read lines for the map. Each line gets checked for only having valid characters and the width, and it can handle multiple linebreak lines (maybe not possible, but just in case)
When wanted amount of lines for the map has been read we start to move the sand.
Transform the string to a char array for easy manipulation. Start from the end. When on an air block, find sand above the spot and if found pull it down. Stops if stone is encountered. Repeate till the second from the top line has been finished reading. Then combine the array and print

Unique errors for the different bad input errors. Exception handling for the "atoi". Handles also 0 sized maps.