chall05
First I check if the wikipedia page exists. After this I search the page starting from the div with the id "mw-content-text" because the page's main content starts there and so it ignores all links before that. I only search for links in <p> tags. It finds all links in those paragraphs and ignores the links with non-ascii values (sets nil in the place). After finding all of the links, skip the ones with nil and skip the ones that have already been visited. Set the url address for the next page to be the first new link found and print the message and check if the next page is Philosophy.