Wikipage hopper
Start by testing an argument is given. Define the base url. Initializing an array which will hold the page history. Make counters for steps taken and page number.

Main-loop: Check if on the right page. Check if visited this page already. Append to history if new. Call for the page reader

Page reader: Parse the page. Test if there is "this page does not exists" element in the page. Find the wanted link-elements using css-selectors and make them into a list.

Back in the main loop: Check if the page contains stuff we want. If no go back a step. If yes start the process again starting from the first link.