Redux
So apparently the bookshelves have sidewalls. So, I made a loop which places the books into the shelves, starting from already used ones. It is not the most thorough, but should work in most cases. Added testes for to fit the largest books in the collection as well as total space available for purchase.