Add krusthol's solution to Bookshelf Problem
Python again, with frequent googling and Learn X in Y cheat sheet and other materials guiding the process. Sprawled off to be messy, I coded this one with a sub-optimal overall structure. But it seems to work, and the standard input seems to be working snappily as well, I made it runnable with the << EOF method, with manual stdin from bash using nothing but the exec itself in terminal, and also | operator seemed to work. Most time used was to build up requested error checking and the described use cases. Less time was thus used to come up with the puzzle logic of sorting the shelves so that you always want to take the biggest shelf into use, and start filling it with the biggest books, thus also sorting the books. Then looping through the possibilities, detecting, if no book could anymore fit into any free shelf to do aborting. I did not give it deep thought, since it solves the example cases in the same minimum amount of shelf units with this fairly simple double iteration logic so I assume it works.