Hardcode hero
Sooo yea...

Basically the solution logic is you reading "square" in the following order: top, right, bottom, left. So I just made a while loop to read thru it all, with a specific exit for even-numbered squares.

Nothing else is really special, print has extra tag to prevent new line. The usage uses arg[0] for the name.

Thought about doing it all in the argv string and counters, but that seemed a more ugly  solution so this was the better alternative