Well yes, but actually no
English grammar rules don't make sense, but still better than the pronounciations.

So, just a simple string looper. Get the string len, have a counter to keep track of positions, sub-function that checks if previous letter was 'c' and next is the third part  of the combo. Inverse the reusult for the "must not" rule.

Default to ret 1 and only change it if the the rule is broken. Could have made it into all one function but somewhat tried to follow the Norm.